refactor: inline ChartInteraction into TimeSeriesChart
